added grid offset to DatasetModel
